decouple registry and compile logic
